requirements.txt
Flask==3.0.3
google-auth==2.34.0
google-cloud-aiplatform==1.70.0
google-genai==0.3.0
